sample_reader.py: overhaul function name lookup
The new implementation assumes that exactly the .text section of each
binary (library or executable) is mapped into memory as described in
the logged memory mappings. Hence, each function address is converted
into an offset into this .text section, and addr2line resolves the address
offset for this section into the actual function.

I'm not sure if this is the correct way, but it appears to work correctly
for libraries and binaries, and also for statically-linked binaries.