samples2calltree.py: better handling of incomplete resolver data
Functions for which no name could be resolved are listed as "_unknown_"
(or "_vdso_" in the special case where it ends in the VDSO, ie. it's
a syscall). If the source file is unknown, a file name is created
that starts with "_in_lib_" followed by the binary file name.