samples2calltree.py: add ugly workaround for http://bugs.kde.org/263594
Kcachegrind crashes if a function has a cost of zero for some event types;
this can easily happen for a multithreaded application when a function is
never ever called by some threads.

As workaround, for multithreaded traces the "zero cost" is now replaced by
a cost of 1, while the actual "1 cost" is replaced by 1000... In the
percentage view in Kcachegrind the small cost vanishes and only the actual
cost remains visible; however the displayed call numbers are off by a
factor of 1000.
Then again, these "call numbers" are not useful values anyway (a sample-
based profiler can only gives percentages by design).