sample_reader.py: fix stack trace addresses to get actual call sites
The stack traces saved by ptrace-sampler actually contain return addresses
rather than addresses of call instructions (well, except for the first
address in each stack trace, which is the actual EIP value).

To get actual call site addresses, the binary is disassembled with objdump,
and the instruction prior to the return target address is used instead.

This of course assumes that a returned call will always continue with the
instruction following the call... Superficial tests support this assumption,
but I'm not sure if advanced compiler optimizations could break this.