sample_reader: revert to using section-relative addressing for addr2line
Apparently separate debug binaries do not necessarily have sections
with same size as the real binary, so we can't use the same offsets
into real and debug binary. Rather, only offsets inside single
sections are identical for both binaries.

So the SymbolResolver now uses addr2line -j again to select a section, but
it also checks which section to use (usually .text, but can also be .plt).
Also, the returned 'offsetInBin' value still refers to the real binary,
so that KCachegrind shows the correct location in its disassembler view.