ptrace-sampler: fix stack trace if eip is in function prolog or epilog
The first few instructions in a function (the "prolog") usually save
previous frame's ebp and then set ebp to current esp. Similarly,
the last few instructions (the "epilog") restore ebp for the calling frame.

If eip is in prolog (or epilog) when taking a sample, the ebp register
value is still (or again) set up for the calling frame; using it to walk
the stack trace would then leave out the immediately calling frame.

To work around this, ptrace-sampler now tries to detect prolog and epilog
(by looking for typical instruction patterns) and outputs an additional
stack trace frame for the calling function if necessary (calculated from
esp rather than ebp).