ptrace-sampler: don't perform run-time disassembling of addresses where we have debug info
The run-time disassembling (for detecting function prolog) is not completely
correct, as it doesn't know if the detected prolog assembler instructions
are actually at begin of function. This can lead to incorrect frames (with
essentially random EIP) being added to backtrace.

I don't know how to fix this correctly; but one workaround to avoid some
of these misdetections is to disable prolog/epilog detection for addresses
where we have debug info anyway.